requirements.txt
fastapi
python-multipart
uvicorn==0.29.0
nest-asyncio
accelerate
scipy
torch
pillow
transformers==4.41.0
edge-tts
groq
requests
pyvips
python-dotenv